requirements.txt
flask==2.2.5
flask-sqlalchemy==3.0.5
flask-restful==0.3.10
flask_prometheus_metrics==1.0.0
flask_jwt_extended==4.6.0
flask-limiter==3.5.0
flask-caching==2.1.0
flasgger==0.9.5
pytest==7.4.0
pytest-playwright==0.4.4
flake8==6.1.0
black==23.7.0
mypy==1.5.1
loguru==0.7.2
myst-parser==2.0.0
pytest-cov==4.1.0
psycopg2==2.9.9
python-dotenv==1.0.0
pandas==2.2.0
openpyxl==3.1.2
pyarrow==15.0.0
SQLAlchemy==2.0.22
requests==2.31.0
Werkzeug==3.0.1
PyJWT==2.8.0
setuptools==69.1.1
isort==5.13.2
